Extracts model class, enum, and top level function into their own files out of DocumentHistoryManager.kt
