+++ app/src/main/java/com/tunepruner/voixt/editor/domain/EditorScreenViewModel.kt
@@ -1,10 +1,11 @@
 package com.tunepruner.voixt.editor.domain
 
 import androidx.lifecycle.ViewModel
+import com.tunepruner.voixt.editor.model.DocumentEvent
 import kotlinx.coroutines.flow.*
 
 class EditorScreenViewModel(editorController: EditorController) : ViewModel() {
-    private val documentHistoryManager = DocumentHistoryManager()
+    private val documentHistoryManager = DocumentHistoryManager(editorController)
     val documentEvents: SharedFlow<DocumentEvent> = documentHistoryManager.documentEvents
 
     suspend fun initialize(words: List<String>) = documentHistoryManager.initialize(words)
