+++ app/src/main/java/com/tunepruner/voixt/editor/model/DocumentEvent.kt
@@ -1,0 +1,7 @@
+package com.tunepruner.voixt.editor.model
+
+class DocumentEvent(
+    val documentEventType: DocumentEventType,
+    val affectedWords: List<String>,
+    val index: Int? = null,
+)
