+++ app/src/main/java/com/tunepruner/voixt/editor/model/DocumentEventType.kt
@@ -1,0 +1,7 @@
+package com.tunepruner.voixt.editor.model
+
+enum class DocumentEventType {
+    ADD,
+    REMOVE,
+    INITIALIZE,
+}
