+++ app/src/main/java/com/tunepruner/voixt/editor/util/StringUtil.kt
@@ -1,0 +1,8 @@
+package com.tunepruner.voixt.editor.util
+
+fun List<String>.concatenateToOneString(): String {
+    val builder = StringBuilder()
+    return fold(builder) { acc, string ->
+        acc.append("$string ")
+    }.toString()
+}
